# auto-remove tempdirs after testrun