--- a/gitbutler-app/src/gb_repository/repository.rs
+++ b/gitbutler-app/src/gb_repository/repository.rs
@@ -520,6 +520,10 @@ impl Repository {
         self.root().join("session")
     }
 
+    pub fn git_repository_path(&self) -> &std::path::Path {
+        self.git_repository.path()
+    }
+
     pub fn session_wd_path(&self) -> std::path::PathBuf {
         self.session_path().join("wd")
     }
@@ -961,38 +965,3 @@ pub enum RemoteError {
     #[error(transparent)]
     Other(#[from] anyhow::Error),
 }
-
-// TODO: this is a unit-test - could use code from `tests::common` via custom module path
-//       to make it work.
-// #[cfg(test)]
-// mod test {
-//     use std::path::PathBuf;
-//
-//     use anyhow::Result;
-//     use pretty_assertions::assert_eq;
-//
-//     use crate::tests::{Case, Suite};
-//
-//     #[test]
-//     fn test_alternates_file_being_set() -> Result<()> {
-//         let Case {
-//             gb_repository,
-//             project_repository,
-//             ..
-//         } = Suite::default().new_case();
-//
-//         let file_content = std::fs::read_to_string(
-//             gb_repository
-//                 .git_repository
-//                 .path()
-//                 .join("objects/info/alternates"),
-//         )?;
-//
-//         let file_content = PathBuf::from(file_content.trim());
-//         let project_path = project_repository.path().to_path_buf().join(".git/objects");
-//
-//         assert_eq!(file_content, project_path);
-//
-//         Ok(())
-//     }
-// }