--- a/gitbutler-app/tests/gb_repository/mod.rs
+++ b/gitbutler-app/tests/gb_repository/mod.rs
@@ -2,6 +2,7 @@ use std::{collections::HashMap, path, thread, time};
 
 use anyhow::Result;
 use pretty_assertions::assert_eq;
+use tempfile::TempDir;
 
 use crate::init_opts_bare;
 use crate::{Case, Suite};
@@ -22,11 +23,12 @@ mod repository {
 
     #[test]
     fn alternates_file_being_set() -> Result<()> {
+        let suite = Suite::default();
         let Case {
             gb_repository,
             project_repository,
             ..
-        } = Suite::default().new_case();
+        } = &suite.new_case();
 
         let file_content = std::fs::read_to_string(
             gb_repository
@@ -43,15 +45,17 @@ mod repository {
     }
 }
 
-fn new_test_remote_repository() -> Result<git2::Repository> {
-    let path = tempfile::tempdir()?.path().to_str().unwrap().to_string();
+fn new_test_remote_repository() -> Result<(git2::Repository, TempDir)> {
+    let tmp = tempfile::tempdir()?;
+    let path = tmp.path().to_str().unwrap().to_string();
     let repo_a = git2::Repository::init_opts(path, &init_opts_bare())?;
-    Ok(repo_a)
+    Ok((repo_a, tmp))
 }
 
 #[test]
 fn get_current_session_writer_should_use_existing_session() -> Result<()> {
-    let Case { gb_repository, .. } = Suite::default().new_case();
+    let suite = Suite::default();
+    let Case { gb_repository, .. } = &suite.new_case();
 
     let current_session_1 = gb_repository.get_or_create_current_session()?;
     let current_session_2 = gb_repository.get_or_create_current_session()?;
@@ -62,7 +66,8 @@ fn get_current_session_writer_should_use_existing_session() -> Result<()> {
 
 #[test]
 fn must_not_return_init_session() -> Result<()> {
-    let Case { gb_repository, .. } = Suite::default().new_case();
+    let suite = Suite::default();
+    let Case { gb_repository, .. } = &suite.new_case();
 
     assert!(gb_repository.get_current_session()?.is_none());
 
@@ -74,13 +79,14 @@ fn must_not_return_init_session() -> Result<()> {
 
 #[test]
 fn must_not_flush_without_current_session() -> Result<()> {
+    let suite = Suite::default();
     let Case {
         gb_repository,
         project_repository,
         ..
-    } = Suite::default().new_case();
+    } = &suite.new_case();
 
-    let session = gb_repository.flush(&project_repository, None)?;
+    let session = gb_repository.flush(project_repository, None)?;
     assert!(session.is_none());
 
     let iter = gb_repository.get_sessions_iterator()?;
@@ -91,30 +97,31 @@ fn must_not_flush_without_current_session() -> Result<()> {
 
 #[test]
 fn non_empty_repository() -> Result<()> {
+    let suite = Suite::default();
     let Case {
         gb_repository,
         project_repository,
         ..
-    } = Suite::default()
-        .new_case_with_files(HashMap::from([(path::PathBuf::from("test.txt"), "test")]));
+    } = &suite.new_case_with_files(HashMap::from([(path::PathBuf::from("test.txt"), "test")]));
 
     gb_repository.get_or_create_current_session()?;
-    gb_repository.flush(&project_repository, None)?;
+    gb_repository.flush(project_repository, None)?;
 
     Ok(())
 }
 
 #[test]
 fn must_flush_current_session() -> Result<()> {
+    let suite = Suite::default();
     let Case {
         gb_repository,
         project_repository,
         ..
-    } = Suite::default().new_case();
+    } = &suite.new_case();
 
     gb_repository.get_or_create_current_session()?;
 
-    let session = gb_repository.flush(&project_repository, None)?;
+    let session = gb_repository.flush(project_repository, None)?;
     assert!(session.is_some());
 
     let iter = gb_repository.get_sessions_iterator()?;
@@ -125,10 +132,11 @@ fn must_flush_current_session() -> Result<()> {
 
 #[test]
 fn list_deltas_from_current_session() -> Result<()> {
-    let Case { gb_repository, .. } = Suite::default().new_case();
+    let suite = Suite::default();
+    let Case { gb_repository, .. } = &suite.new_case();
 
     let current_session = gb_repository.get_or_create_current_session()?;
-    let writer = deltas::Writer::new(&gb_repository)?;
+    let writer = deltas::Writer::new(gb_repository)?;
     writer.write(
         "test.txt",
         &vec![deltas::Delta {
@@ -137,7 +145,7 @@ fn list_deltas_from_current_session() -> Result<()> {
         }],
     )?;
 
-    let session_reader = sessions::Reader::open(&gb_repository, &current_session)?;
+    let session_reader = sessions::Reader::open(gb_repository, &current_session)?;
     let deltas_reader = deltas::Reader::new(&session_reader);
     let deltas = deltas_reader.read(None)?;
 
@@ -156,13 +164,14 @@ fn list_deltas_from_current_session() -> Result<()> {
 
 #[test]
 fn list_deltas_from_flushed_session() {
+    let suite = Suite::default();
     let Case {
         gb_repository,
         project_repository,
         ..
-    } = Suite::default().new_case();
+    } = &suite.new_case();
 
-    let writer = deltas::Writer::new(&gb_repository).unwrap();
+    let writer = deltas::Writer::new(gb_repository).unwrap();
     writer
         .write(
             "test.txt",
@@ -172,9 +181,9 @@ fn list_deltas_from_flushed_session() {
             }],
         )
         .unwrap();
-    let session = gb_repository.flush(&project_repository, None).unwrap();
+    let session = gb_repository.flush(project_repository, None).unwrap();
 
-    let session_reader = sessions::Reader::open(&gb_repository, &session.unwrap()).unwrap();
+    let session_reader = sessions::Reader::open(gb_repository, &session.unwrap()).unwrap();
     let deltas_reader = deltas::Reader::new(&session_reader);
     let deltas = deltas_reader.read(None).unwrap();
 
@@ -191,13 +200,14 @@ fn list_deltas_from_flushed_session() {
 
 #[test]
 fn list_files_from_current_session() {
-    let Case { gb_repository, .. } = Suite::default().new_case_with_files(HashMap::from([(
+    let suite = Suite::default();
+    let Case { gb_repository, .. } = &suite.new_case_with_files(HashMap::from([(
         path::PathBuf::from("test.txt"),
         "Hello World",
     )]));
 
     let current = gb_repository.get_or_create_current_session().unwrap();
-    let reader = sessions::Reader::open(&gb_repository, &current).unwrap();
+    let reader = sessions::Reader::open(gb_repository, &current).unwrap();
     let files = reader.files(None).unwrap();
 
     assert_eq!(files.len(), 1);
@@ -209,21 +219,22 @@ fn list_files_from_current_session() {
 
 #[test]
 fn list_files_from_flushed_session() {
+    let suite = Suite::default();
     let Case {
         gb_repository,
         project_repository,
         ..
-    } = Suite::default().new_case_with_files(HashMap::from([(
+    } = &suite.new_case_with_files(HashMap::from([(
         path::PathBuf::from("test.txt"),
         "Hello World",
     )]));
 
     gb_repository.get_or_create_current_session().unwrap();
     let session = gb_repository
-        .flush(&project_repository, None)
+        .flush(project_repository, None)
         .unwrap()
         .unwrap();
-    let reader = sessions::Reader::open(&gb_repository, &session).unwrap();
+    let reader = sessions::Reader::open(gb_repository, &session).unwrap();
     let files = reader.files(None).unwrap();
 
     assert_eq!(files.len(), 1);
@@ -236,7 +247,7 @@ fn list_files_from_flushed_session() {
 #[tokio::test]
 async fn remote_syncronization() {
     // first, crate a remote, pretending it's a cloud
-    let cloud = new_test_remote_repository().unwrap();
+    let (cloud, _tmp) = new_test_remote_repository().unwrap();
     let api_project = ApiProject {
         name: "test-sync".to_string(),
         description: None,
@@ -331,7 +342,7 @@ async fn remote_syncronization() {
 #[tokio::test]
 async fn remote_sync_order() {
     // first, crate a remote, pretending it's a cloud
-    let cloud = new_test_remote_repository().unwrap();
+    let (cloud, _tmp) = new_test_remote_repository().unwrap();
     let api_project = projects::ApiProject {
         name: "test-sync".to_string(),
         description: None,
@@ -453,11 +464,12 @@ async fn remote_sync_order() {
 
 #[test]
 fn gitbutler_file() {
+    let suite = Suite::default();
     let Case {
         gb_repository,
         project_repository,
         ..
-    } = Suite::default().new_case();
+    } = &suite.new_case();
 
     let session = gb_repository.get_or_create_current_session().unwrap();