--- a/gitbutler-app/tests/git/config.rs
+++ b/gitbutler-app/tests/git/config.rs
@@ -2,7 +2,7 @@ use crate::test_repository;
 
 #[test]
 pub fn set_str() {
-    let repo = test_repository();
+    let (repo, _tmp) = test_repository();
     let mut config = repo.config().unwrap();
     config.set_str("test.key", "test.value").unwrap();
     assert_eq!(
@@ -13,7 +13,7 @@ pub fn set_str() {
 
 #[test]
 pub fn set_bool() {
-    let repo = test_repository();
+    let (repo, _tmp) = test_repository();
     let mut config = repo.config().unwrap();
     config.set_bool("test.key", true).unwrap();
     assert!(config.get_bool("test.key").unwrap().unwrap());
@@ -21,14 +21,14 @@ pub fn set_bool() {
 
 #[test]
 pub fn get_string_none() {
-    let repo = test_repository();
+    let (repo, _tmp) = test_repository();
     let config = repo.config().unwrap();
     assert_eq!(config.get_string("test.key").unwrap(), None);
 }
 
 #[test]
 pub fn get_bool_none() {
-    let repo = test_repository();
+    let (repo, _tmp) = test_repository();
     let config = repo.config().unwrap();
     assert_eq!(config.get_bool("test.key").unwrap(), None);
 }