--- a/gitbutler-app/tests/sessions/database.rs
+++ b/gitbutler-app/tests/sessions/database.rs
@@ -4,7 +4,7 @@ use gitbutler_app::sessions::{session, Database, Session, SessionId};
 
 #[test]
 fn insert_query() -> anyhow::Result<()> {
-    let db = test_database();
+    let (db, _tmp) = test_database();
     println!("0");
     let database = Database::new(db);
     println!("1");
@@ -47,7 +47,7 @@ fn insert_query() -> anyhow::Result<()> {
 
 #[test]
 fn update() -> anyhow::Result<()> {
-    let db = test_database();
+    let (db, _tmp) = test_database();
     let database = Database::new(db);
 
     let project_id = ProjectId::generate();